# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Project PM API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/milestones:
    get:
      operationId: listMilestones
      tags: [milestones]
      summary: List milestones for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of milestones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Milestone"
    post:
      operationId: createMilestone
      tags: [milestones]
      summary: Create a milestone
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MilestoneInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Milestone"
  /milestones/{id}:
    patch:
      operationId: updateMilestone
      tags: [milestones]
      summary: Update a milestone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MilestoneUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Milestone"
    delete:
      operationId: deleteMilestone
      tags: [milestones]
      summary: Delete a milestone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Tasks ─────────────────────────────────────────────────────────────────
...
  /members:
    get:
      operationId: listMembers
      tags: [members]
      summary: List all team members
      responses:
        "200":
          description: Team members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"
    post:
      operationId: createMember
      tags: [members]
      summary: Create a team member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /members/{id}:
    patch:
      operationId: updateMember
      tags: [members]
      summary: Update a team member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
    delete:
      operationId: deleteMember
      tags: [members]
      summary: Delete a team member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Project Members ──────────────────────────────────────────────────────
...
  /folders:
    get:
      operationId: listFolders
      tags: [folders]
      summary: List document folders (optionally filtered by project)
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
        - name: global
          in: query
          description: If true, return only global (non-project) folders
          schema:
            type: boolean
        - name: taskId
          in: query
          description: Filter to folders attached to a specific task
          schema:
            type: integer
        - name: includeTasks
          in: query
          description: When combined with projectId, also include folders attached to any task within that project.
          schema:
            type: boolean
      responses:
        "200":
          description: Folders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocumentFolder"
    post:
      operationId: createFolder
      tags: [folders]
      summary: Create a document folder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentFolderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentFolder"
  /folders/{id}:
    patch:
      operationId: updateFolder
      tags: [folders]
      summary: Rename or move a folder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentFolderUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentFolder"
    delete:
      operationId: deleteFolder
      tags: [folders]
      summary: Delete an empty folder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "409":
          description: Folder is not empty
...
components:
  schemas:
...
    Milestone:
      type: object
      required: [id, projectId, name, status, createdAt]
      properties:
        id:          { type: integer }
        projectId:    { type: integer }
        name:        { type: string }
        description: { type: ["string", "null"] }
        dueDate:     { type: ["string", "null"] }
        status:      { type: string, enum: [pending, in_progress, completed, missed] }
        color:       { type: ["string", "null"] }
        createdAt:   { type: string }
    MilestoneInput:
      type: object
      required: [name]
      properties:
        name:        { type: string, minLength: 1 }
        description: { type: string }
        dueDate:     { type: string }
        status:      { type: string, enum: [pending, in_progress, completed, missed] }
        color:       { type: string }
    MilestoneUpdate:
      type: object
      properties:
        name:        { type: string }
        description: { type: string }
        dueDate:     { type: string }
        status:      { type: string, enum: [pending, in_progress, completed, missed] }
        color:       { type: string }

    # ── Task ────────────────────────────────────────────────────────────────
...
    Member:
      type: object
      required: [id, name, role, createdAt]
      properties:
        id:           { type: integer }
        name:         { type: string }
        email:        { type: ["string", "null"] }
        role:         { type: string, enum: [producer, engineer, it, integrator, manager, contractor] }
        department:   { type: ["string", "null"] }
        avatarUrl:    { type: ["string", "null"] }
        title:        { type: ["string", "null"], description: "Job title, e.g. 'Senior AV Engineer'." }
        phone:        { type: ["string", "null"], description: "Office or main phone." }
        mobilePhone:  { type: ["string", "null"] }
        location:     { type: ["string", "null"], description: "City, timezone, or site assignment." }
        company:      { type: ["string", "null"], description: "External company, for contractors/integrators." }
        notes:        { type: ["string", "null"] }
        createdAt:    { type: string }
    MemberInput:
      type: object
      required: [name, role]
      properties:
        name:         { type: string, minLength: 1 }
        email:        { type: string }
        role:         { type: string, enum: [producer, engineer, it, integrator, manager, contractor] }
        department:   { type: string }
        avatarUrl:    { type: string }
        title:        { type: string }
        phone:        { type: string }
        mobilePhone:  { type: string }
        location:     { type: string }
        company:      { type: string }
        notes:        { type: string }
    MemberUpdate:
      type: object
      description: "Partial update. Send `null` for an optional field to clear it; omit to leave unchanged."
      properties:
        name:         { type: string }
        role:         { type: string, enum: [producer, engineer, it, integrator, manager, contractor] }
        email:        { type: ["string", "null"] }
        department:   { type: ["string", "null"] }
        avatarUrl:    { type: ["string", "null"] }
        title:        { type: ["string", "null"] }
        phone:        { type: ["string", "null"] }
        mobilePhone:  { type: ["string", "null"] }
        location:     { type: ["string", "null"] }
        company:      { type: ["string", "null"] }
        notes:        { type: ["string", "null"] }

    # ── Project Member ──────────────────────────────────────────────────────
...
    DocumentFolder:
      type: object
      required: [id, name, createdAt]
      properties:
        id:        { type: integer }
        projectId: { type: ["integer", "null"] }
        taskId:    { type: ["integer", "null"] }
        parentId:  { type: ["integer", "null"] }
        name:      { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
    DocumentFolderInput:
      type: object
      required: [name]
      properties:
        projectId: { type: integer }
        taskId:    { type: integer }
        parentId:  { type: integer }
        name:      { type: string, minLength: 1 }
    DocumentFolderUpdate:
      type: object
      properties:
        name:     { type: string, minLength: 1 }
        parentId: { type: ["integer", "null"] }
...